+++ src/main/java/cn/memoryzy/json/actions/child/DonateAction.java
@@ -1,8 +1,7 @@
 package cn.memoryzy.json.actions.child;
 
 import cn.memoryzy.json.bundles.JsonAssistantBundle;
-import cn.memoryzy.json.constant.HyperLinks;
-import com.intellij.ide.BrowserUtil;
+import cn.memoryzy.json.ui.SupportDialog;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.DumbAwareAction;
 import icons.JsonAssistantIcons;
@@ -14,22 +13,13 @@ import org.jetbrains.annotations.NotNull;
  */
 public class DonateAction extends DumbAwareAction {
 
-    private final String hyperLink;
-
-    public DonateAction(String text, String hyperLink) {
+    public DonateAction(String text) {
         super(text, JsonAssistantBundle.messageOnSystem("action.donate.description"), JsonAssistantIcons.DONATE);
-        this.hyperLink = hyperLink;
     }
 
     @Override
     public void actionPerformed(@NotNull AnActionEvent e) {
-        if (HyperLinks.reachableAtomic.get()) {
-            BrowserUtil.browse(hyperLink);
-        } else {
-            // 构建弹窗
-
-
-        }
+        new SupportDialog().show();
     }
 
 }