+++ src/main/java/cn/memoryzy/json/listener/HyperLinkListenerImpl.java
@@ -1,24 +1,19 @@
 package cn.memoryzy.json.listener;
 
-import cn.hutool.core.net.url.UrlBuilder;
 import cn.memoryzy.json.bundles.JsonAssistantBundle;
 import cn.memoryzy.json.constant.HyperLinks;
-import cn.memoryzy.json.ui.JsonToJavaBeanDialog;
 import cn.memoryzy.json.utils.PlatformUtil;
 import com.intellij.ide.BrowserUtil;
 import com.intellij.openapi.diagnostic.Logger;
-import com.intellij.openapi.ui.MessageType;
 import com.intellij.openapi.ui.popup.Balloon;
 import com.intellij.openapi.ui.popup.JBPopupFactory;
 import com.intellij.ui.HyperlinkAdapter;
 import com.intellij.ui.awt.RelativePoint;
+import com.intellij.util.ui.JBUI;
 
 import javax.swing.event.HyperlinkEvent;
 import java.awt.*;
-import java.io.IOException;
 import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.Objects;
 
 /**
  * @author Memory
@@ -30,15 +25,20 @@ public class HyperLinkListenerImpl extends HyperlinkAdapter {
 
     @Override
     protected void hyperlinkActivated(HyperlinkEvent e) {
-        String description = e.getDescription();
-        if (Objects.equals(HyperLinks.PLUGIN_SHARE_LINK, description)) {
-            execShareLinkAction(e);
-        } else if (Objects.equals(HyperLinks.PLUGIN_EMAIL_LINK, description)){
-            execEmailLinkAction(e);
-        } else {
-            BrowserUtil.browse(description);
+        String url = e.getDescription();
+        switch (url) {
+            case HyperLinks.PLUGIN_SHARE_LINK:
+                execShareLinkAction(e);
+                break;
+            case HyperLinks.PLUGIN_EMAIL_LINK:
+                execEmailLinkAction(e);
+                break;
+            default:
+                BrowserUtil.browse(url);
+                break;
         }
     }
+
 
     private void execEmailLinkAction(HyperlinkEvent e) {
         String uri = "mailto:" + HyperLinks.EMAIL_LINK;
@@ -53,8 +53,16 @@ public class HyperLinkListenerImpl extends HyperlinkAdapter {
         PlatformUtil.setClipboard(HyperLinks.MARKETPLACE_LINK);
         Component component = e.getInputEvent().getComponent();
         JBPopupFactory.getInstance()
-                .createHtmlTextBalloonBuilder(JsonAssistantBundle.messageOnSystem("dialog.support.share.text"), MessageType.INFO, null)
+                .createHtmlTextBalloonBuilder(JsonAssistantBundle.messageOnSystem("dialog.support.share.text"),
+                        null,
+                        JBUI.CurrentTheme.NotificationInfo.backgroundColor(),
+                        null)
                 .setShadow(true)
+                .setHideOnAction(true)
+                .setHideOnClickOutside(true)
+                .setHideOnFrameResize(true)
+                .setHideOnKeyOutside(true)
+                .setHideOnLinkClick(true)
                 .setFadeoutTime(5000L)
                 .createBalloon()
                 .show(new RelativePoint(component, new Point(component.getWidth() / 4, component.getHeight())), Balloon.Position.below);