+++ src/main/java/cn/memoryzy/json/ui/basic/MyUI.java
@@ -1,0 +1,15 @@
+package cn.memoryzy.json.ui.basic;
+
+import com.intellij.openapi.Disposable;
+
+/**
+ * @author Memory
+ * @since 2024/7/30
+ */
+public class MyUI implements Disposable {
+
+    @Override
+    public void dispose() {
+
+    }
+}